Safety device for blinds
It is an Arduino based device which can be fixed to blind person’s stick
to make them alert when anything is close to them. It gives out buzzer
sound when the danger is very near to them. Ultrasonic sensor was
interfaced with arduino, giving the exact distance from any object.
Ultrasonic sensor sends out ultrasonic waves which get reflected back
from the object infornt. This travel time is calculated and distance is
measured.